@@ -4,42 +4,46 @@
     <groupId>org.reactome.server.tools</groupId>
     <artifactId>analysis-service</artifactId>
     <packaging>war</packaging>
-    <version>3.0.2</version>
+    <version>3.1.0-SNAPSHOT</version>
     <name>analysis-service</name>
     <description>Pathway enrichment and expression analysis as well as species comparison tool</description>
 
     <properties>
         <deploymentPath>/</deploymentPath>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <jdk.version>11</jdk.version>
+        <spring.version>5.3.7</spring.version>
 
-        <analysis.core>3.3.7-SNAPSHOT</analysis.core>
-        <analysis.report>1.1.6-SNAPSHOT</analysis.report>
-        <spring.version>4.3.10.RELEASE</spring.version>
-        <reactome.utils.version>1.0.10</reactome.utils.version>
+        <analysis.core>3.4.0-SNAPSHOT</analysis.core>
+        <analysis.report>1.2.0-SNAPSHOT</analysis.report>
+        <reactome.utils.version>1.1.0-SNAPSHOT</reactome.utils.version>
 
-        <sl4j.version>1.7.2</sl4j.version>
-        <fileupload.version>1.3</fileupload.version>
-        <commons.codec>1.11</commons.codec>
-        <jackson.version>2.9.8</jackson.version>
-        <swagger.version>2.7.0</swagger.version>
-        <tika.version>1.12</tika.version>
-        <jackson.asl.version>1.9.12</jackson.asl.version>
+        <fileupload.version>1.4</fileupload.version>
+        <commons.codec>1.15</commons.codec>
+        <jackson.version>2.12.2</jackson.version>
+        <swagger.version>3.0.0</swagger.version>
+        <tika.version>2.0.0-ALPHA</tika.version>
 
-        <servlet.version>4.0.0-b01</servlet.version>
-        <jsp.api.version>2.1</jsp.api.version>
+        <servlet.version>4.0.1</servlet.version>
+        <jsp.api.version>2.3.3</jsp.api.version>
         <jstl.api.version>1.2</jstl.api.version>
-        <xml.security.version>1.0.4</xml.security.version>
-        <xalan.version>2.7.1</xalan.version>
-        <junit.version>4.13</junit.version>
-        <mail.version>1.6.0</mail.version>
+        <xalan.version>2.7.2</xalan.version>
+        <junit.version>5.8.0-M1</junit.version>
+        <mail.version>1.6.2</mail.version>
         <activation.version>1.1.1</activation.version>
-
-        <maven.compiler.version>3.5.1</maven.compiler.version>
-        <jdk.version>1.8</jdk.version>
-
+        <jaxb.core.version>3.0.0</jaxb.core.version>
+        <commons.io.version>2.8.0</commons.io.version>
+        <slf4j.api.version>2.0.0-alpha1</slf4j.api.version>
+        <slf4j.log4j12.version>2.0.0-alpha1</slf4j.log4j12.version>
         <hamcrest.library.version>2.2</hamcrest.library.version>
-        <json.path.version>2.4.0</json.path.version>
+        <json.path.version>2.5.0</json.path.version>
+        <javax.annotation.version>1.3.2</javax.annotation.version>
 
+        <maven.compiler.version>3.8.1</maven.compiler.version>
+        <tomcat7.maven.version>2.2</tomcat7.maven.version>
+        <jetty.maven.version>8.1.14.v20131031</jetty.maven.version>
+        <maven.surefire.version>3.0.0-M5</maven.surefire.version>
+        <maven.jar.version>3.2.0</maven.jar.version>
     </properties>
 
     <developers>
