@@ -1,14 +1,14 @@
 package org.reactome.server.analysis.service.controller;
 
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.reactome.server.analysis.AppTests;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
@@ -17,13 +17,13 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 
-@RunWith(SpringJUnit4ClassRunner.class)
+@ExtendWith(SpringExtension.class)
 @ContextConfiguration({"file:src/test/resources/mvc-dispatcher-servlet-test.xml"})
 @WebAppConfiguration
 public class ImporterControllerTest extends AppTests {
 
-    @Before
-    public void prepare(){
+    @BeforeEach
+    public void prepare() {
         generateToken("P02452 P08123 P02461 P12110 P49674 P35222 P09668 Q9NQC7");
     }
 
@@ -33,14 +33,11 @@ public class ImporterControllerTest extends AppTests {
         String url = String.format("/download/%s/result.json", AppTests.token);
         MvcResult result = mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, null);
         String responseResult = result.getResponse().getContentAsString();
-
         mockMvcPostResult("/import/", responseResult);
-
     }
 
     @Test
     public void getPostFile() throws Exception {
-
         // generate a content string
         String url = String.format("/download/%s/result.json", AppTests.token);
         MvcResult result = mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, null);
@@ -49,7 +46,7 @@ public class ImporterControllerTest extends AppTests {
         MockMultipartFile importFile = new MockMultipartFile("file", "result.json", "multipart/form-data", response.getBytes());
 
         MockHttpServletRequestBuilder builder =
-                MockMvcRequestBuilders.fileUpload("/import/form")
+                MockMvcRequestBuilders.multipart("/import/form")
                         .file(importFile);
 
         this.getMockMvc().perform(builder)
