@@ -1,11 +1,11 @@
 package org.reactome.server.analysis.service.controller;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.reactome.server.analysis.AppTests;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.request.MockMultipartHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
@@ -13,10 +13,11 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.fileUpload;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringJUnit4ClassRunner.class)
+
+@ExtendWith(SpringExtension.class)
 @ContextConfiguration({"file:src/test/resources/mvc-dispatcher-servlet-test.xml"})
 @WebAppConfiguration
 public class MappingControllerTest extends AppTests {
@@ -70,7 +71,7 @@ public class MappingControllerTest extends AppTests {
         files.add(overrepresentationFile);
         files.add(expressionFile);
 
-        MockMultipartHttpServletRequestBuilder requestBuilder = fileUpload("/mapping/form/projection");
+        MockMultipartHttpServletRequestBuilder requestBuilder = multipart("/mapping/form/projection");
 
         for (MockMultipartFile file : files) {
             requestBuilder.file(file);
@@ -89,7 +90,7 @@ public class MappingControllerTest extends AppTests {
         files.add(overrepresentationFile);
         files.add(expressionFile);
 
-        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.fileUpload("/mapping/form");
+        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.multipart("/mapping/form");
 
         for (MockMultipartFile file : files) {
             requestBuilder.file(file);
