@@ -6,6 +6,7 @@ import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;
 import springfox.documentation.service.Contact;
+import springfox.documentation.service.Tag;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
@@ -26,7 +27,17 @@ public class AnalysisServiceSwaggerConfig {
                 .select()
                 .apis(RequestHandlerSelectors.any())
                 .paths(PathSelectors.any())
-                .build().apiInfo(apiInfo());
+                .build()
+                .tags(new Tag("database", "Database info queries"))
+                .tags(new Tag("download", "Methods to download different views of a result"))
+                .tags(new Tag("identifier", "Queries for only one identifier"))
+                .tags(new Tag("identifiers",  "Queries for multiple identifiers"))
+                .tags(new Tag("import", "Imports an external result"))
+                .tags(new Tag("mapping", "Identifiers mapping methods"))
+                .tags(new Tag("report", "Retrieves report files in PDF format"))
+                .tags(new Tag("species",  "Species comparison"))
+                .tags(new Tag( "token", "Previous queries filter"))
+                .apiInfo(apiInfo());
     }
 
     private ApiInfo apiInfo() {
