@@ -166,7 +166,7 @@ public class HeaderFooterCacher extends Thread {
     // trusting all certificate
     @SuppressWarnings("Duplicates")
     private void doTrustToCertificates() throws NoSuchAlgorithmException, KeyManagementException {
-        Security.addProvider(new com.sun.net.ssl.internal.ssl.Provider());
+        Security.addProvider(Security.getProvider("SunJSSE"));
         TrustManager[] trustAllCerts = new TrustManager[]{
                 new X509TrustManager() {
                     public X509Certificate[] getAcceptedIssuers() {
